requirements.txt
# Core libraries for environment management
python-dotenv

# Libraries for Google Sheets API
gspread
google-auth-oauthlib
